+++ cedar-policy/src/api.rs
@@ -647,6 +647,14 @@ impl Entities {
     pub fn write_to_json(&self, f: impl std::io::Write) -> std::result::Result<(), EntitiesError> {
         self.0.write_to_json(f)
     }
+
+    #[doc = include_str!("../experimental_warning.md")]
+    /// Visualize an `Entities` object in the graphviz `dot`
+    /// format. Entity visualization is best-effort and not well tested.
+    /// Feel free to submit an issue if you are using this feature and would like it improved.
+    pub fn to_dot_str(&self) -> String {
+        self.0.to_dot_str()
+    }
 }
 
 /// Utilities for defining `IntoIterator` over `Entities`